Migrate SQS tests to a dedicated file: sqs.test.ts
